добавлены обои и дороботон скрипт

+++ install.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+GO_VERSION="1.26.3"
+FreeCAD_URL="https://release-assets.githubusercontent.com/github-production-release-asset/5736080/d97e16b8-8010-4118-b489-e908208bb313?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-05-16T08%3A29%3A11Z&rscd=attachment%3B+filename%3DFreeCAD_1.1.1-Linux-x86_64-py311.AppImage&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-05-16T07%3A28%3A47Z&ske=2026-05-16T08%3A29%3A11Z&sks=b&skv=2018-11-09&sig=zzNPSSeifbA6AYcqfzBiEGIKl8mYJ8bRjre%2Fn%2BcSYQQ%3D&jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmVsZWFzZS1hc3NldHMuZ2l0aHVidXNlcmNvbnRlbnQuY29tIiwia2V5Ijoia2V5MSIsImV4cCI6MTc3ODkyMDU0MiwibmJmIjoxNzc4OTE2OTQyLCJwYXRoIjoicmVsZWFzZWFzc2V0cHJvZHVjdGlvbi5ibG9iLmNvcmUud2luZG93cy5uZXQifQ.hLLQfXw85gbd2XjXTGK5F0y9fKuBoU0fSAK54WZAagM&response-content-disposition=attachment%3B%20filename%3DFreeCAD_1.1.1-Linux-x86_64-py311.AppImage&response-content-type=application%2Foctet-stream"
+PineconeMC_URL="https://release-assets.githubusercontent.com/github-production-release-asset/561202233/4299d4a1-d383-4d7c-a151-4c571be0180b?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-05-16T08%3A28%3A29Z&rscd=attachment%3B+filename%3DPineconeMC-Linux-x86_64.AppImage&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-05-16T07%3A27%3A44Z&ske=2026-05-16T08%3A28%3A29Z&sks=b&skv=2018-11-09&sig=ulQObc4JZ2Q%2BZefs1dZGszufgVrAVaTWydS0smt%2Bo9I%3D&jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmVsZWFzZS1hc3NldHMuZ2l0aHVidXNlcmNvbnRlbnQuY29tIiwia2V5Ijoia2V5MSIsImV4cCI6MTc3ODkyMDMwMCwibmJmIjoxNzc4OTE2NzAwLCJwYXRoIjoicmVsZWFzZWFzc2V0cHJvZHVjdGlvbi5ibG9iLmNvcmUud2luZG93cy5uZXQifQ.JnD5foWRlH5KzYr66bJrA-8OIQDjOGPMoHChXJx_INw&response-content-disposition=attachment%3B%20filename%3DPineconeMC-Linux-x86_64.AppImage&response-content-type=application%2Foctet-stream"
+
 exec > >(tee -i log.txt) 2>&1
 
 set -e
@@ -77,13 +81,20 @@ typeset -U path
 EOF
 
 echo "=== 4. Установка Go ==="
-GO_VERSION="1.26.3"
-sudo rm -rf /usr/local/go
 
-wget https://go.dev/dl/go${GO_VERSION}.linux-amd64.tar.gz -P /tmp
-sudo tar -C /usr/local -xzf /tmp/go${GO_VERSION}.linux-amd64.tar.gz
-rm /tmp/go${GO_VERSION}.linux-amd64.tar.gz
-echo 'export PATH=$PATH:/usr/local/go/bin' >> ~/.bashrc
+if [ -x /usr/local/go/bin/go ] && /usr/local/go/bin/go version | grep -q "go${GO_VERSION}"; then
+    echo "Go версии ${GO_VERSION} уже установлен. Пропускаем скачивание."
+else
+    echo "Установка Go версии ${GO_VERSION}..."
+    sudo rm -rf /usr/local/go
+    wget https://go.dev/dl/go${GO_VERSION}.linux-amd64.tar.gz -P /tmp
+    sudo tar -C /usr/local -xzf /tmp/go${GO_VERSION}.linux-amd64.tar.gz
+    rm /tmp/go${GO_VERSION}.linux-amd64.tar.gz
+    if ! grep -q '/usr/local/go/bin' ~/.bashrc; then
+        echo 'export PATH=$PATH:/usr/local/go/bin' >> ~/.bashrc
+    fi
+fi
+
 /usr/local/go/bin/go version
 
 echo "=== 5. Установка VS Code ==="
@@ -93,6 +104,7 @@ sudo apt update
 sudo apt install -y code
 
 echo "=== 6. Установка конфигов ==="
+flatpak remote-add --if-not-exists flathub https://dl.flathub.org/repo/flathub.flatpakrepo
 git config --global user.name "LiyaFPV"
 git config --global user.email "lbkmzc942@gmail.com"
 git config --global init.defaultBranch main
@@ -101,9 +113,11 @@ git config --global --list
 mkdir -p ~/.config
 mkdir -p ~/Applications
 mkdir -p ~/userScripts
+mkdir -p ~/Wallpapers
 cp -r Applications/* ~/Applications/ 2>/dev/null || true
 cp -r userScripts/* ~/userScripts/ 2>/dev/null || true
 cp -r config/* ~/.config/ 2>/dev/null || true
+cp -r Wallpapers/* ~/Wallpapers/ 2>/dev/null || true
 systemctl --user --now enable wireplumber.service
 if [ -f ~/userScripts/install.sh ]; then
     chmod +x ~/userScripts/install.sh
@@ -129,21 +143,20 @@ echo 'KERNEL=="ttyACM*", MODE="0666"' | sudo tee -a /etc/udev/rules.d/99-usb-ser
 sudo udevadm control --reload-rules && sudo udevadm trigger
 
 echo "=== 8. Установка Telegram и Discord (Нативные пакеты) ==="
-
-# 1. Установка Telegram через PPA
 sudo apt update
-sudo apt install -y telegram-desktop
+sudo apt install telegram-desktop -y
 echo "Скачивание Discord..."
 wget -O /tmp/discord.deb "https://discord.com/api/download?platform=linux&format=deb"
 echo "Установка Discord..."
-sudo apt install -y /tmp/discord.deb
+wget -O /tmp/discord.deb "https://discord.com/api/download?platform=linux&format=deb"
+sudo apt install /tmp/discord.deb -y
 rm /tmp/discord.deb
 
 echo "=== 9. скачевание AppImage ==="
-wget -O ~/Applications/FreeCAD.AppImage "https://release-assets.githubusercontent.com/github-production-release-asset/5736080/d97e16b8-8010-4118-b489-e908208bb313?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-05-16T08%3A29%3A11Z&rscd=attachment%3B+filename%3DFreeCAD_1.1.1-Linux-x86_64-py311.AppImage&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-05-16T07%3A28%3A47Z&ske=2026-05-16T08%3A29%3A11Z&sks=b&skv=2018-11-09&sig=zzNPSSeifbA6AYcqfzBiEGIKl8mYJ8bRjre%2Fn%2BcSYQQ%3D&jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmVsZWFzZS1hc3NldHMuZ2l0aHVidXNlcmNvbnRlbnQuY29tIiwia2V5Ijoia2V5MSIsImV4cCI6MTc3ODkyMDU0MiwibmJmIjoxNzc4OTE2OTQyLCJwYXRoIjoicmVsZWFzZWFzc2V0cHJvZHVjdGlvbi5ibG9iLmNvcmUud2luZG93cy5uZXQifQ.hLLQfXw85gbd2XjXTGK5F0y9fKuBoU0fSAK54WZAagM&response-content-disposition=attachment%3B%20filename%3DFreeCAD_1.1.1-Linux-x86_64-py311.AppImage&response-content-type=application%2Foctet-stream" \ 
+wget -O ~/Applications/FreeCAD.AppImage  ${FreeCAD_URL} \ 
 && chmod +x ~/Applications/FreeCAD.AppImage
 
-wget -O ~/Applications/PineconeMC-Linux-x86_64.AppImage "https://release-assets.githubusercontent.com/github-production-release-asset/561202233/4299d4a1-d383-4d7c-a151-4c571be0180b?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-05-16T08%3A28%3A29Z&rscd=attachment%3B+filename%3DPineconeMC-Linux-x86_64.AppImage&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-05-16T07%3A27%3A44Z&ske=2026-05-16T08%3A28%3A29Z&sks=b&skv=2018-11-09&sig=ulQObc4JZ2Q%2BZefs1dZGszufgVrAVaTWydS0smt%2Bo9I%3D&jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmVsZWFzZS1hc3NldHMuZ2l0aHVidXNlcmNvbnRlbnQuY29tIiwia2V5Ijoia2V5MSIsImV4cCI6MTc3ODkyMDMwMCwibmJmIjoxNzc4OTE2NzAwLCJwYXRoIjoicmVsZWFzZWFzc2V0cHJvZHVjdGlvbi5ibG9iLmNvcmUud2luZG93cy5uZXQifQ.JnD5foWRlH5KzYr66bJrA-8OIQDjOGPMoHChXJx_INw&response-content-disposition=attachment%3B%20filename%3DPineconeMC-Linux-x86_64.AppImage&response-content-type=application%2Foctet-stream"
+wget -O ~/Applications/PineconeMC-Linux-x86_64.AppImage  ${PineconeMC_URL} \
 && chmod +x ~/Applications/PineconeMC-Linux-x86_64.AppImage
 
 echo "=== 10. Смена оболочки на Zsh ==="
